requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-WTF==1.2.1
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pymongo==4.7.2
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2
